feat: Implement search screen and enhanced review features
- Introduce dedicated search screen with history and popular keywords.
- Enhance review screens (AddReviewScreen, AddReviewToProductScreen, EditReviewScreen) with improved UI, styling, and form validation.
- Update rating system to support and store half-star ratings (double) in Supabase.
- Refactor Review model and related controllers/screens to handle double ratings.
- Update README.md to reflect schema changes and new features.
- Update unit tests for AddReviewScreen to match new UI texts.
- Add search_controller.dart for future search state management (currently untracked).

+++ lib/presentation/providers/add_review_controller.dart
@@ -188,7 +188,7 @@ class AddReviewController extends StateNotifier<AddReviewState> {
         userId: user.id,
         productId: newProduct.id,
         reviewText: state.reviewText,
-        rating: state.rating.toInt(),
+        rating: state.rating,
       );
 
       await reviewRepository.createReview(newReview);
